requirements.txt
cycler==0.10.0
decorator==4.4.2
deskew==0.10.3
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pkg-resources==0.0.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
